Debug first generation models, move BVC to a separate utils file
